Added unit death and trying to add an actual match

+++ backend/engine/entity.go
@@ -2,7 +2,7 @@ package engine
 
 //Entity currently useless, just an index
 type Entity struct {
-	PlayerTag byte // 0 or 1, player 1 or 2
+	PlayerTag int // 0 or 1, player 1 or 2
 	Index     int
 	Name      string
 	Active    bool
@@ -16,5 +16,5 @@ type EntityData struct {
 	State    string `json:"state"`
 	Index    int    `json:"index"`
 	//	Path     []Vector `json:"path"`
-	Tag byte `json:"player_tag"`
+	Tag int `json:"player_tag"`
 }
